requirements.txt
google-api-python-client==1.7.11
oauth2client==4.1.3
google-auth-oauthlib==0.4.1
beautifulsoup4==4.8.1